requirements.txt
openai-agents
fastapi
uvicorn[standard]
python-dotenv
